moved css to themes directory
