+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@designcise/next-theme-toggle": "^4.0.0",
     "@netlify/plugin-nextjs": "^5.11.4",
+    "@tailwindcss/cli": "^4.1.12",
     "@wojtekmaj/react-hooks": "^1.21.0",
     "gray-matter": "^4.0.3",
     "lucide-react": "^0.511.0",
@@ -23,6 +24,7 @@
     "next": "^15.3.2",
     "next-burger-menu": "^1.1.0",
     "next-intl": "^4.1.0",
+    "next-themes": "^0.4.6",
     "pnpm": "^10.11.0",
     "process": "^0.11.10",
     "react": "^19.0.0",
@@ -45,6 +47,7 @@
     "@eslint/eslintrc": "^3",
     "@next/eslint-plugin-next": "^15.3.3",
     "@nextui-org/react": "^2.6.11",
+    "@tailwindcss/postcss": "^4.1.12",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-burger-menu": "^2.8.7",
@@ -57,8 +60,10 @@
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.2.0",
     "mini-css-extract-plugin": "^2.9.2",
+    "postcss": "^8.5.6",
     "react-burger-menu": "^3.1.0",
     "rehype-highlight": "^7.0.2",
+    "tailwindcss": "^4.1.12",
     "trim-newlines": "^5.0.0",
     "typescript": "^5"
   }
