+++ src/app/[locale]/documents/page.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
-import styles from "../page.module.css";
-import '../fonts.css';
+import styles from "../themes/page.module.css";
+import '../themes/fonts.css';
 import { useTranslations } from "next-intl";
