+++ src/app/[locale]/themes/page.module.css
@@ -1,10 +1,12 @@
 @font-face {
   font-family: 'spaceGrotesk';
-  src: url('../../fonts/SpaceGrotesk[wght].woff2');
+  src: url('../../../fonts/SpaceGrotesk[wght].woff2');
 }
 
-.page {
-  --button-primary-hover: #ccc;
+/* @import "global.css"; */
+
+[data-theme='light'] .page {
+  /* --button-primary-hover: #ccc;
   --button-secondary-hover: #d4472d;
   --background-colour: #accab2;
   --secondary-colour: #d4472d;
@@ -16,7 +18,7 @@
   --minor_inverting_filter_hover: invert(93%) sepia(11%) saturate(5067%) hue-rotate(319deg) brightness(96%) contrast(90%);
   --minor_filter: brightness(0) saturate(100%) invert(77%) sepia(42%) saturate(657%) hue-rotate(332deg) brightness(97%) contrast(94%);
   --minor_filter_hover: invert(89%) sepia(100%) saturate(46%) hue-rotate(326deg) brightness(111%) contrast(111%);
-  --white_filter: invert(100%) sepia(100%) saturate(0%) hue-rotate(187deg) brightness(105%) contrast(101%);
+  --white_filter: invert(100%) sepia(100%) saturate(0%) hue-rotate(187deg) brightness(105%) contrast(101%); */
 
   display: grid;
   grid-template-rows: 20px 1fr 20px;
@@ -29,7 +31,7 @@
 }
 
 .page.dark {
-  --background-colour: #060f37;
+  /* --background-colour: #060f37;
   --secondary-colour: #e9a752;
   --tertiary-colour: #1E6496;
   --quaternary-colour: #d4472d;
@@ -42,7 +44,7 @@
   --minor_inverting_filter_hover: invert(93%) sepia(11%) saturate(5067%) hue-rotate(319deg) brightness(96%) contrast(90%);
   --minor_filter: brightness(0) saturate(100%) invert(77%) sepia(42%) saturate(657%) hue-rotate(332deg) brightness(97%) contrast(94%);
   --minor_filter_hover: invert(89%) sepia(100%) saturate(46%) hue-rotate(326deg) brightness(111%) contrast(111%);
-  --white_filter: invert(100%) sepia(100%) saturate(0%) hue-rotate(187deg) brightness(105%) contrast(101%);
+  --white_filter: invert(100%) sepia(100%) saturate(0%) hue-rotate(187deg) brightness(105%) contrast(101%); */
 
   display: grid;
   grid-template-rows: 20px 1fr 20px;
