+++ src/app/components/ToggleLocaleButton/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 import Image from 'next/image';
 
-import styles from "../../../app/[locale]/page.module.css";
+import styles from "../../../app/[locale]/themes/page.module.css";
 
 import { useLocale, useTranslations } from 'next-intl';
